requirements.txt
asgiref==3.6.0
Django==4.1.7
django-environ==0.9.0
djangorestframework==3.14.0
marshmallow==3.19.0
mysqlclient==2.1.1
packaging==23.0
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
